CSSTUDIO-1978: fix header component, update export structure.
